Persisting notifications to db

+++ src/main/java/com/sourcery/gymapp/backend/events/LikeNotificationEvent.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 public record LikeNotificationEvent(
         UUID ownerId,
         UUID routineId,
-        long likesCount
+        int likesCount
 ) {
 }
