+++ src/main/resources/db/changelog/changes/userprofile/v2.0-routine-like-notification-changelog.sql
@@ -10,6 +10,6 @@ CREATE TABLE IF NOT EXISTS user_profiles.like_notification
     modified_by          UUID                        NOT NULL,
     owner_id             UUID                        NOT NULL,
     routine_id           UUID                        NOT NULL,
-    likes_count          BIGINT                      NOT NULL,
+    likes_count          INTEGER                     NOT NULL,
     CONSTRAINT pk_like_notification PRIMARY KEY (id)
 );
